Entregado

+++ assignments/10Videojuego/src/exercise.py
@@ -1,6 +1,6 @@
 def main():
-    a = float(input("Dame la cantidad de juegos nuevos: "))
-    b = float(input("Dame la cantidad de juegos usados: "))
+    a = int(input("Dame la cantidad de juegos nuevos: "))
+    b = int(input("Dame la cantidad de juegos usados: "))
     c = (a*1000)+(b*350)
     print("El total de la compra es:",c)
